(PC-2555): BSR -> moved Booking & BookingCancellation to layout folder
