+++ src/components/layout/Booking/BookingForm/BookingFormContent/__specs__/BookingFormContent.spec.jsx
