+++ src/components/layout/Booking/BookingForm/BookingFormContent/__specs__/BookingFormContentContainer.spec.js
