+++ src/components/layout/Booking/BookingForm/decorators/utils/__specs__/onCalendarUpdate.spec.js
