+++ src/components/layout/Booking/BookingHeader/__specs__/BookingHeader.spec.jsx
