+++ src/components/layout/Booking/BookingSuccess/__specs__/BookingSuccess.spec.jsx
