+++ src/components/layout/Booking/__specs__/Booking.spec.jsx
