+++ src/components/layout/Booking/__specs__/BookingContainer.spec.js
