+++ src/components/layout/Booking/utils/__specs__/flattenErrors.spec.js
