+++ src/components/layout/Booking/utils/__specs__/getCalendarProvider.spec.js
