+++ src/components/layout/Booking/utils/__specs__/handleRedirect.spec.js
