+++ src/components/layout/Booking/utils/__specs__/parseHoursByStockId.spec.js
