+++ src/components/layout/Booking/utils/handleRedirect.js
