+++ src/components/layout/Booking/utils/parseHoursByStockId.js
