+++ src/components/pages/discovery/Discovery.jsx
@@ -3,8 +3,8 @@ import React, { Fragment, PureComponent } from 'react'
 import { Route } from 'react-router-dom'
 
 import DeckContainer from './Deck/DeckContainer'
-import BookingContainer from '../Booking/BookingContainer'
-import BookingCancellationContainer from '../BookingCancellation/BookingCancellationContainer'
+import BookingContainer from '../../layout/Booking/BookingContainer'
+import BookingCancellationContainer from '../../layout/BookingCancellation/BookingCancellationContainer'
 import AbsoluteFooterContainer from '../../layout/AbsoluteFooter/AbsoluteFooterContainer'
 import LoaderContainer from '../../layout/Loader/LoaderContainer'
 import isDetailsView from '../../../helpers/isDetailsView'
